Move nugget to a separate file
